seperated into new components and fixed some issues
