+++ client/src/app/category/category.module.ts
@@ -1,17 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CategoryCardComponent } from './components/category-card/category-card.component';
-import {
-  CategoryDialogComponent,
-  EditConfirmationComponent,
-} from './components/category-dialog/category-dialog.component';
-import {
-  CategoryComponent,
-  CategoryDeleteConfirmComponent,
-} from './components/category/category.component';
+import { CategoryDialogComponent } from './components/category-dialog/category-dialog.component';
+
 import { AuthGuard } from '../auth/auth.guard';
 import { RouterModule, Routes } from '@angular/router';
 import { SharedModule } from '../shared/shared.module';
+import { CategoryComponent } from './components/category/category.component';
+import { CategoryDeleteConfirmComponent } from './components/category-delete-confirm/category-delete-confirm.component';
+import { EditConfirmationComponent } from './components/edit-confirm/edit-confirm.component';
 
 const routes: Routes = [
   {
