+++ client/src/app/category/components/category-delete-confirm/category-delete-confirm.component.scss
@@ -1,0 +1,34 @@
+@use "src/sass/mixins.scss" as mixin;
+@use "src/sass/variables.scss" as var;
+
+.delete-confirm {
+  display: flex;
+  height: 100%;
+  flex-direction: column;
+  &__header {
+    @include mixin.flex-between;
+
+    p {
+      margin: 0;
+      font-weight: bold;
+    }
+  }
+
+  &__body {
+    flex: 1;
+    margin-top: 20px;
+  }
+}
+
+::ng-deep .snackbar-success {
+  background-color: var.$successBg;
+  color: var.$textLight !important;
+  button {
+    color: var.$textLight !important;
+  }
+}
+
+::ng-deep .snackbar-error {
+  background-color: var.$warningBg;
+  color: var.$textDanger !important;
+}
