+++ client/src/app/category/components/category-dialog/category-dialog.component.html
@@ -6,16 +6,6 @@
 </div>
 <mat-dialog-content class="dialog-body">
   <form id="categoryForm" [formGroup]="addCategoryForm" (ngSubmit)="onSubmit()">
-    <!-- <mat-form-field class="dialog-body__row" appearance="outline">
-      <mat-label>Category type</mat-label>
-      <mat-select formControlName="type" cdkFocusInitial>
-        <mat-option [value]="'income'"> Income </mat-option>
-        <mat-option [value]="'expense'"> Expense </mat-option>
-      </mat-select>
-      <mat-error *ngIf="addCategoryForm.invalid">{{
-        getTypeError()
-      }}</mat-error>
-    </mat-form-field> -->
     <div class="dialog-body__actions">
       <button
         type="button"
