+++ client/src/app/category/components/edit-confirm/edit-confirm.component.scss
@@ -1,0 +1,19 @@
+@use "src/sass/mixins.scss" as mixin;
+.edit-confirm {
+  display: flex;
+  height: 100%;
+  flex-direction: column;
+  &__header {
+    @include mixin.flex-between;
+
+    p {
+      margin: 0;
+      font-weight: bold;
+    }
+  }
+
+  &__body {
+    flex: 1;
+    margin-top: 20px;
+  }
+}
