+++ client/src/app/category/components/edit-confirm/edit-confirmation.component.html
@@ -11,8 +11,8 @@
   </div>
   <div mat-dialog-conten class="edit-confirm__body">
     <p>
-      Are you sure you want to change the title of category? All your incomes
-      will be updated according to the renamed category
+      Are you sure you want to change the title of category? All your
+      transactions will be updated according to the renamed category
     </p>
   </div>
   <mat-dialog-actions class="dialog-body__actions" align="end">
