+++ client/src/app/transaction/components/transaction-dialog/transaction-dialog.component.scss
@@ -68,25 +68,6 @@
   }
 }
 
-.cancel-confirm {
-  display: flex;
-  height: 100%;
-  flex-direction: column;
-  &__header {
-    @include mixin.flex-between;
-
-    p {
-      margin: 0;
-      font-weight: bold;
-    }
-  }
-
-  &__body {
-    flex: 1;
-    margin-top: 20px;
-  }
-}
-
 ::ng-deep .snackbar-success {
   background-color: var.$successBg;
   color: var.$textLight !important;
